Sound manager was added. Algorithm that opens cells in 'zero' directions added. Minor improvements.

+++ JapaneseMinesweeper/UIColor+ColorFromHexString.h
@@ -11,6 +11,6 @@
 @interface UIColor (ColorFromHexString)
 
 + (UIColor *)colorFromHexString: (NSString *)hexString;
-+ (UIColor *)colorFromInteger: (NSUInteger)integer;
++ (UIColor *)colorFromInteger: (UInt32)integer;
 
 @end
